Changed health/damage system to be more conducive to damage boosts in powerups. Added smaller health bars for slimes. (keep?)

+++ core/src/com/upa/gun/Player.java
@@ -18,7 +18,7 @@ public class Player extends Entity {
 
     Vector2 spawnPoint;
 
-    double bulletCooldown;
+    float bulletCooldown;
 
     public boolean topStop = false;
     public boolean botStop = false;
